requirements.txt
fastapi
uvicorn[standard]
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
langgraph
langsmith
chromadb
pypdf
python-dotenvpip
langchain-tavily
sqlalchemy
pymysq1
psycop2-binary
pandas
matplotlib
faker
chromadb
supabase
psycopg2-binary